[step4] 라우팅 시 정적 페이지 요청 먼저 처리하고 동적 요청 처리하도록 변경
